move agent detection utils to a separate module
