+++ js/src/modules/off-screen-menu.js
@@ -1,3 +1,7 @@
+// Some general utils for user agent detection.
+// Could potentially be replaced by modernizr in future if it gets out of control
+var agentDetection = require('../utils/agent-detection')($);
+
 // Always export a function which takes jQuery as an argument
 // This ensures each module is testable
 module.exports = function($) {
@@ -28,12 +32,12 @@ module.exports = function($) {
                 initOffScreenIos();
             },
             androidJsSupport: function() {
-                if (detectAndroidStockBrowser()) {
+                if (agentDetection.detectAndroidStockBrowser()) {
                     return true;
                 }
             },
             iOSJsSupport: function() {
-                if (iOSversion() < 5) {
+                if (agentDetection.iOSversion() < 5) {
                     return true;
                 }
             }
@@ -64,18 +68,18 @@ module.exports = function($) {
             });
         }
 
-        function iOSversion() {
-            if (/iP(hone|od touch|ad)/.test(navigator.platform)) {
-                var v = (navigator.appVersion).match(/OS (\d+)_(\d+)_?(\d+)?/);
-                return parseInt(v[1], 10);
-            }
-        }
+        // function iOSversion() {
+        //     if (/iP(hone|od touch|ad)/.test(navigator.platform)) {
+        //         var v = (navigator.appVersion).match(/OS (\d+)_(\d+)_?(\d+)?/);
+        //         return parseInt(v[1], 10);
+        //     }
+        // }
 
-        function detectAndroidStockBrowser() {
-            var nua = navigator.userAgent;
-            var isAndroid = ((nua.indexOf('Mozilla/5.0') > -1 && nua.indexOf('Android ') > -1 && nua.indexOf('AppleWebKit') > -1) && !(nua.indexOf('Chrome') > -1));
-            return isAndroid;
-        }
+        // function detectAndroidStockBrowser() {
+        //     var nua = navigator.userAgent;
+        //     var isAndroid = ((nua.indexOf('Mozilla/5.0') > -1 && nua.indexOf('Android ') > -1 && nua.indexOf('AppleWebKit') > -1) && !(nua.indexOf('Chrome') > -1));
+        //     return isAndroid;
+        // }
 
         // Initialise off screen menu js for browsers that don't support the CSS only method
         if (api.androidJsSupport()) {
